requirements.txt
Pillow
requests
imagehash
argparse
torch 
opencv-python
torchvision
